Add Security Checks
Add Security Checks

+++ PowerToolbox/Services/Download/DownloadSchedulerService.cs
@@ -352,6 +352,11 @@ namespace PowerToolbox.Services.Download
         /// </summary>
         internal static void CreateDownload(string fileLink, string filePath)
         {
+            if (string.IsNullOrEmpty(fileLink) || string.IsNullOrEmpty(filePath))
+            {
+                return;
+            }
+
             if (string.Equals(doEngineMode, DownloadOptionsService.DoEngineModeList[0]))
             {
                 DeliveryOptimizationService.CreateDownload(fileLink, filePath);
@@ -371,6 +376,11 @@ namespace PowerToolbox.Services.Download
         /// </summary>
         internal static void ContinueDownload(string downloadID)
         {
+            if (string.IsNullOrEmpty(downloadID))
+            {
+                return;
+            }
+
             if (string.Equals(doEngineMode, DownloadOptionsService.DoEngineModeList[0]))
             {
                 DeliveryOptimizationService.ContinueDownload(downloadID);
@@ -390,6 +400,11 @@ namespace PowerToolbox.Services.Download
         /// </summary>
         internal static void PauseDownload(string downloadID)
         {
+            if (string.IsNullOrEmpty(downloadID))
+            {
+                return;
+            }
+
             if (string.Equals(doEngineMode, DownloadOptionsService.DoEngineModeList[0]))
             {
                 DeliveryOptimizationService.PauseDownload(downloadID);
@@ -409,6 +424,11 @@ namespace PowerToolbox.Services.Download
         /// </summary>
         internal static void DeleteDownload(string downloadID)
         {
+            if (string.IsNullOrEmpty(downloadID))
+            {
+                return;
+            }
+
             if (string.Equals(doEngineMode, DownloadOptionsService.DoEngineModeList[0]))
             {
                 DeliveryOptimizationService.DeleteDownload(downloadID);
